refactor: update import paths for person and news schemas to server-actions directory
